requirements.txt
beautifulsoup4==4.12.0
psycopg2_binary==2.9.6
Requests==2.29.0
